Actualizacion de bug y renombramiento de clases
