+++ app/src/main/java/com/example/koffer/MainActivity.java
@@ -1,115 +1,15 @@
 package com.example.koffer;
 
-import android.support.constraint.ConstraintLayout;
-import android.support.v4.view.ViewPager;
+import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.text.Html;
-import android.view.View;
-import android.widget.Button;
-import android.widget.LinearLayout;
-import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
-
-    private ViewPager mSlideViewPager;
-    private LinearLayout mDotLayout;
-
-    private SlideAdapter slideAdapter;
-    private TextView[] mDots;
-
-    private Button mNextBtn;
-    private Button mPrevBtn;
-    private int mCurrentPage;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         setTheme(R.style.AppTheme);
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        mSlideViewPager = (ViewPager) findViewById(R.id.slideViewPager);
-        mDotLayout = (LinearLayout) findViewById(R.id.dotsLayout);
-
-        mNextBtn = (Button) findViewById(R.id.nextBtn);
-        mPrevBtn = (Button) findViewById(R.id.prevBtn);
-
-        slideAdapter = new SlideAdapter(this);
-
-        mSlideViewPager.setAdapter(slideAdapter);
-
-        addDotsIndicator(0);
-        mSlideViewPager.addOnPageChangeListener(viewListener);
-
-        mNextBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mSlideViewPager.setCurrentItem(mCurrentPage + 1);
-            }
-        });
-
-        mPrevBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mSlideViewPager.setCurrentItem(mCurrentPage - 1);
-            }
-        });
     }
-
-    public void addDotsIndicator(int position){
-        mDots = new TextView[3];
-        mDotLayout.removeAllViews();
-
-        for (int i = 0; i < mDots.length; i++) {
-            mDots[i] = new TextView(this);
-            mDots[i].setText(Html.fromHtml("&#8226"));
-            mDots[i].setTextSize(35);
-            mDots[i].setTextColor(getResources().getColor(R.color.colorTransparentWhite));
-            mDotLayout.addView(mDots[i]);
-        }
-
-        if (mDots.length > 0) {
-            mDots[position].setTextColor(getResources().getColor(R.color.colorWhite));
-        }
-    }
-
-    ViewPager.OnPageChangeListener viewListener = new  ViewPager.OnPageChangeListener() {
-
-        @Override
-        public void onPageScrolled(int i, float v, int i1) {
-
-        }
-
-        @Override
-        public void onPageSelected(int i) {
-            addDotsIndicator(i);
-
-            mCurrentPage = i;
-
-            if (i == 0) {
-                mNextBtn.setEnabled(true);
-                mPrevBtn.setEnabled(false);
-                mPrevBtn.setVisibility(View.INVISIBLE);
-                mNextBtn.setText("SIGUIENTE");
-                mPrevBtn.setText("");
-            }else if (i == mDots.length - 1) {
-                mNextBtn.setEnabled(true);
-                mPrevBtn.setEnabled(true);
-                mPrevBtn.setVisibility(View.VISIBLE);
-                mNextBtn.setText("ACABAR");
-                mPrevBtn.setText("ATRÁS");
-            } else {
-                mNextBtn.setEnabled(true);
-                mPrevBtn.setEnabled(true);
-                mPrevBtn.setVisibility(View.VISIBLE);
-                mNextBtn.setText("SIGUIENTE");
-                mPrevBtn.setText("ATRÁS");
-            }
-        }
-
-        @Override
-        public void onPageScrollStateChanged(int i) {
-
-        }
-    };
 }
