+++ app/src/main/java/com/example/koffer/SlideActivity.java
@@ -1,0 +1,128 @@
+package com.example.koffer;
+
+import android.content.Intent;
+import android.support.v4.view.ViewPager;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.text.Html;
+import android.view.View;
+import android.widget.Button;
+import android.widget.LinearLayout;
+import android.widget.TextView;
+
+public class SlideActivity extends AppCompatActivity {
+    private ViewPager mSlideViewPager;
+    private LinearLayout mDotLayout;
+
+    private SlideAdapter slideAdapter;
+    private TextView[] mDots;
+
+    private Button mNextBtn;
+    private Button mPrevBtn;
+    private Button mFinishBtn;
+    private int mCurrentPage;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_slide);
+
+        mSlideViewPager = (ViewPager) findViewById(R.id.slideViewPager);
+        mDotLayout = (LinearLayout) findViewById(R.id.dotsLayout);
+
+        mNextBtn = (Button) findViewById(R.id.nextBtn);
+        mPrevBtn = (Button) findViewById(R.id.prevBtn);
+        mFinishBtn = (Button) findViewById(R.id.finishBtn);
+
+        slideAdapter = new SlideAdapter(this);
+
+        mSlideViewPager.setAdapter(slideAdapter);
+
+        addDotsIndicator(0);
+        mSlideViewPager.addOnPageChangeListener(viewListener);
+
+        mNextBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                mSlideViewPager.setCurrentItem(mCurrentPage + 1);
+            }
+        });
+
+        mPrevBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                mSlideViewPager.setCurrentItem(mCurrentPage - 1);
+            }
+        });
+
+        mFinishBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(SlideActivity.this, Main2Activity.class);
+                startActivity(intent);
+            }
+        });
+    }
+
+    public void addDotsIndicator(int position){
+        mDots = new TextView[3];
+        mDotLayout.removeAllViews();
+
+        for (int i = 0; i < mDots.length; i++) {
+            mDots[i] = new TextView(this);
+            mDots[i].setText(Html.fromHtml("&#8226"));
+            mDots[i].setTextSize(35);
+            mDots[i].setTextColor(getResources().getColor(R.color.colorTransparentWhite));
+            mDotLayout.addView(mDots[i]);
+        }
+
+        if (mDots.length > 0) {
+            mDots[position].setTextColor(getResources().getColor(R.color.colorWhite));
+        }
+    }
+
+    ViewPager.OnPageChangeListener viewListener = new  ViewPager.OnPageChangeListener() {
+
+        @Override
+        public void onPageScrolled(int i, float v, int i1) {
+
+        }
+
+        @Override
+        public void onPageSelected(int i) {
+            addDotsIndicator(i);
+
+            mCurrentPage = i;
+
+            if (i == 0) {
+                mNextBtn.setEnabled(true);
+                mPrevBtn.setEnabled(false);
+                mFinishBtn.setEnabled(false);
+                mFinishBtn.setVisibility(View.INVISIBLE);
+                mPrevBtn.setVisibility(View.INVISIBLE);
+                mNextBtn.setVisibility(View.VISIBLE);
+            }else if (i == mDots.length - 1) {
+                mNextBtn.setEnabled(false);
+                mPrevBtn.setEnabled(true);
+                mFinishBtn.setEnabled(true);
+                mNextBtn.setVisibility(View.INVISIBLE);
+                mPrevBtn.setVisibility(View.VISIBLE);
+                mFinishBtn.setVisibility(View.VISIBLE);
+
+
+            } else {
+                mNextBtn.setEnabled(true);
+                mPrevBtn.setEnabled(true);
+                mFinishBtn.setEnabled(false);
+                mFinishBtn.setVisibility(View.INVISIBLE);
+                mPrevBtn.setVisibility(View.VISIBLE);
+                mNextBtn.setVisibility(View.VISIBLE);
+            }
+        }
+
+        @Override
+        public void onPageScrollStateChanged(int i) {
+
+        }
+    };
+}
